refactor: COM model and pipeline management (#10)
* refactor(d3d11): let device always own immediate context

* refactor: create certain objects with initial refcount 1

This is inconsistent, normally we expect refcount 0 for a new object. However this leads to some probelm, for example, you can't really query interfaces inside constructor, otherwise you either get a leak or unexpected destruction.

It might be feasible to move all other com objects creation to this semantic.

* refactor(d3d11): d3d11 state object cache

They are all effectively owned by D3D11Device despite the object refcount. This matches the behavior that declared on msdn.

* chore: use dxmt::thread instead of std counterpart

* fix(d3d11): various tweak for robust pipeline cache lookup

And of course at the expense of memory leak that should be gracefully handled later

* refactor(d3d11): don't hold ref to pipeline state objects in context

* chore(d3d11): rename `transferFrom` to `transfer`